Moved more stuff into service.
