+++ app/frontend/src/hooks/useChatService.ts
@@ -1,14 +1,100 @@
 import { useMemo } from "react"
 import { PersistenceUtils } from "../util/persistence";
 import { useChatStore } from "../stores/ChatStore";
+import { useAppStore } from "../stores/AppStore";
+import { useTranslation } from "react-i18next";
+import { SNACKBAR_DEBOUNCE_KEYS } from "../constants";
+import { isACompletion } from "../utils";
 
 export const useChatService = () => {
+    const { t } = useTranslation()
     const chatStore = useChatStore();
-    const { currentChatIndex, setChatIndexToLoadOrDelete, setChatHistoriesDescriptions, setDefaultChatHistory, setCurrentChatHistory, setCurrentChatIndex } = useChatStore();
+    const snackbars = useAppStore((state) => state.snackbars);
+    const { currentChatIndex, chatHistoriesDescriptions, setChatIndexToLoadOrDelete, setChatHistoriesDescriptions, setDefaultChatHistory, setCurrentChatHistory, setCurrentChatIndex: chatStoreSetCurrentChatIndex } = useChatStore();
     // This is a custom hook that provides chat-related services. We use useMemo to
     // memoize the value of the service to avoid unnecessary re-renders.
+
+    const setCurrentChatIndex = (index: number) => {
+        // Set the index in local storage
+        PersistenceUtils.setCurrentChatIndex(index);
+        // Update the state
+        chatStoreSetCurrentChatIndex(index);
+    }
+
+    const saveChatHistories = (updatedChatHistory: ChatHistory) => {
+        try {
+            const chatHistories = PersistenceUtils.getChatHistories();
+            chatHistories[currentChatIndex] = updatedChatHistory;
+            PersistenceUtils.setChatHistories(chatHistories);
+        } catch (error) {
+            if (
+                error instanceof DOMException &&
+                error.name === "QuotaExceededError"
+            ) {
+                console.error("LocalStorage is full:", error);
+                snackbars.show(
+                    t("storage.full"),
+                    SNACKBAR_DEBOUNCE_KEYS.STORAGE_FULL_ERROR
+                );
+            }
+            console.error("Failed to save to localStorage:", error);
+        }
+    }
+
+    const renameChat = (newDescription: string, indexToUpdate: number) => {
+        const chatHistories = PersistenceUtils.getChatHistories();
+        const updatedChatHistories = [...chatHistories];
+        const updatedChatHistory: ChatHistory = {
+            ...chatHistories[indexToUpdate],
+            description: newDescription,
+        };
+        updatedChatHistories[indexToUpdate] = updatedChatHistory;
+        PersistenceUtils.setChatHistories(updatedChatHistories);
+        if (currentChatIndex === indexToUpdate) {
+            setCurrentChatHistory(updatedChatHistory);
+        }
+        setChatHistoriesDescriptions(
+            updatedChatHistories.map(
+                (chatHistory, index) =>
+                    chatHistory.description || "Conversation " + (index + 1)
+            )
+        );
+    };
+
+    const handleLoadSavedChat = (index: number) => {
+        const chatHistories = PersistenceUtils.getChatHistories();
+        if (chatHistories) {
+            const newChat = chatHistories[index];
+            setCurrentChatHistory(newChat);
+            setCurrentChatIndex(index);
+        }
+    };
+
+    const handleNewChat = () => {
+        const chatHistories = PersistenceUtils.getChatHistories();
+        const newChatIndex = chatHistoriesDescriptions.length;
+        if (chatHistories.length === 10 || newChatIndex >= 10) {
+            snackbars.show(
+                t("chat.history.full"),
+                SNACKBAR_DEBOUNCE_KEYS.CHAT_HISTORY_FULL_ERROR
+            );
+        } else {
+            setCurrentChatIndex(newChatIndex);
+            setDefaultChatHistory()
+            setChatHistoriesDescriptions([
+                ...chatHistoriesDescriptions,
+                "Conversation " + (chatHistoriesDescriptions.length + 1),
+            ]);
+        }
+    };
+
     const memoized = useMemo(() => {
         return {
+            setCurrentChatIndex,
+            saveChatHistories,
+            renameChat,
+            handleLoadSavedChat,
+            handleNewChat,
             deleteSavedChat: async (chatIndexToLoadOrDelete: number) => {
                 const chatHistories = PersistenceUtils.getChatHistories();
                 const updatedChatHistories = [
@@ -41,9 +127,38 @@ export const useChatService = () => {
 
                 setChatIndexToLoadOrDelete(null);
                 PersistenceUtils.setChatHistories(updatedChatHistories);
-            }
+            },
+            updateLastMessage(message_chunk: string) {
+                setCurrentChatHistory((prevChatHistory) => {
+                    const updatedChatItems = prevChatHistory?.chatItems.map(
+                        (item, itemIndex) => {
+                            if (
+                                itemIndex === prevChatHistory.chatItems.length - 1 &&
+                                isACompletion(item)
+                            ) {
+                                return {
+                                    ...item,
+                                    message: {
+                                        ...item.message,
+                                        content: message_chunk,
+                                    },
+                                };
+                            }
+                            return item;
+                        }
+                    );
+
+                    const updatedChatHistory: ChatHistory = {
+                        ...prevChatHistory,
+                        chatItems: updatedChatItems,
+                    };
+
+                    saveChatHistories(updatedChatHistory);
+                    return updatedChatHistory;
+                });
+            },
         }
-    }, [chatStore])
+    }, [chatStore, snackbars])
 
     return memoized
 }
